Move component to Advanced and prettier format fix
