+++ packages/admin-ui/src/pages/system/components/Profile/index.tsx
@@ -3,22 +3,12 @@ import Card from "components/Card";
 import SubTitle from "components/SubTitle";
 import { ChangePassword } from "./ChangePassword";
 import { SignOut } from "./SignOut";
-import { ChangeDappnodeWebName } from "./ChangeDappnodeWebName";
 
 type ShowStatus = "loading" | "show" | "hide";
 
 export default function Profile() {
   return (
     <>
-      <SubTitle>Change DappNode Name</SubTitle>
-      <Card spacing>
-        <div>
-        Dappnode name only affects to the User Interface. It's the name appears
-        on the top navigation bar.
-        </div>
-        <ChangeDappnodeWebName />
-      </Card>
-
       <SubTitle>Change UI password</SubTitle>
       <Card spacing>
         <div>
